Reorganize files and add missing public API export for asyncData
