+++ libs/ngx-nuts-and-bolts/src/index.ts
@@ -4,16 +4,19 @@ export * from './lib/animations/height.animation';
 
 // Testing utilities
 export * from './lib/testing/extract-public/extract-public.type';
+export * from './lib/testing/async-data/async-data.helper';
 
 // Table state helpers
-export * from './lib/helpers/table-state/table-state.helper';
+export * from './lib/utilities/table-state/table-state';
 
 // Directives
 export * from './lib/directives/in-view/in-view.module';
 export * from './lib/directives/in-view/in-view.directive';
 export * from './lib/directives/in-view/in-view.testing.module';
 export * from './lib/directives/in-view/in-view.testing.directive';
-export * from './lib/directives/loading-state/loading-state';
+
+// Utilities
+export * from './lib/utilities/loading-state/loading-state';
 
 // Pipes
 export * from './lib/pipes/enum-property/enum-property.module';
